+++ src/IPC/horizontal_lines.csv
@@ -1,3 +1,3 @@
--996,-864,-744,-636,-546
+-720,-318,-258,-204,-150
 -1.5708,-1.5708,-1.5708,-1.5708,-1.5708
-5,6,7,8,9
+1,6,7,8,9
