+++ src/IPC/vertical_lines.csv
@@ -1,3 +1,3 @@
-749,837,1001,1165,1231
-0.17453,0.12217,0,-0.10472,-0.17453
-3,4,6,8,9
+343,407,528,586,643
+0.087266,0.034907,-0.069813,-0.12217,-0.17453
+4,5,7,8,9
